comments made through responder now used to check if feedback is needed
We now ascertain whether feedback is required by filtering through
comments created by the VortexFeedback account to see if the user had
already responded to any developer comments. This will ensure that users
that damage/modify/delete their application state will not be asked for
their response if they've already responded in the past.

Also fixed an issue causing the responder to remain stuck in the fetch
comments placeholder state.